+++ src/components/Waypoints.jsx
@@ -34,13 +34,15 @@ const propTypes = {
       opacity: PropTypes.number,
       weight: PropTypes.number,
     }))
-  })
+  }),
+  onSummaryCalculated: PropTypes.func,
 };
 
 const defaultProps = {
   testId: 'component',
   fitRoutes: true,
   summaryTemplate: '',
+  waypoints: null,
   lineOptions: {
     styles: [
       {color: '#65803c', opacity: 0.15, weight: 7},
@@ -48,14 +50,16 @@ const defaultProps = {
       {color: '#8ebf42', opacity: 1, weight: 2},
     ],
   },
+  onSummaryCalculated: null,
 };
 
-export const Waypoints = ({
-                            testId,
-                            waypoints,
-                            summaryTemplate,
-                            lineOptions,
-                            fitRoutes,
+export const Waypoints = React.memo(({
+  testId,
+  waypoints,
+  summaryTemplate,
+  lineOptions,
+  fitRoutes,
+  onSummaryCalculated,
 }) => {
   const context = useLeafletContext();
 
@@ -114,6 +118,12 @@ export const Waypoints = ({
   React.useEffect(() => {
     const container = context.layerContainer || context.map;
 
+    control.on('routesfound', (e) => {
+      const {routes} = e;
+      const {summary} = routes[0];
+      onSummaryCalculated && onSummaryCalculated(summary);
+    });
+
     container.addControl(control);
 
     return () => {
@@ -122,7 +132,7 @@ export const Waypoints = ({
   });
 
   return null;
-};
+});
 
 Waypoints.propTypes = propTypes;
 Waypoints.defaultProps = defaultProps;